do some more code cleanups
