+++ bin/lib/__init__.py
@@ -1,2 +1,2 @@
 from .keys import OMDB_API_KEY, TMDB_API_KEY
-from .get_movie_from_args import get_movie
+from .main import get_movie
