+++ bin/lib/posters.py
